+++ src/native/common/xcc_fmt.h
@@ -31,8 +31,8 @@
 extern "C" {
 #endif
 
-int xcc_fmt_snprintf(char *buffer, size_t buffer_size, const char *format, ...);
-int xcc_fmt_vsnprintf(char *buffer, size_t buffer_size, const char *format, va_list args);
+size_t xcc_fmt_snprintf(char *buffer, size_t buffer_size, const char *format, ...);
+size_t xcc_fmt_vsnprintf(char *buffer, size_t buffer_size, const char *format, va_list args);
 
 #ifdef __cplusplus
 }
